+++ automation/src/epilot/models/shared/getexecutionsresp.py
@@ -11,7 +11,6 @@ from epilot import utils
 
 @dataclasses.dataclass
 class GetExecutionsResp:
-    r"""List of automation executions"""
     results: list[shared_automationexecution.AutomationExecution] = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('results') }})
     total: float = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('total') }})