+++ automation/src/epilot/models/shared/startexecutionrequest.py
@@ -11,7 +11,6 @@ from typing import Optional
 
 @dataclasses.dataclass
 class StartExecutionRequestInput:
-    r"""Execution parameters"""
     entity_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('entity_id'), 'exclude': lambda f: f is None }})